Fix metadata for early issues
The metadata for the early issues of the R Journal (2009-2012) were
slightly screwed up (only for non-linked articles, e.g., in News and Notes),
as pointed out by Kurt in an email dated 2014/10/28. Mostly the metadata was
all there, but under wrong issues. Hopefully this commit fixes these errors.

We may additionally want to make these link to PDF snippets and add them to the
bibliography, as in issues from 2013 onward. But this will need more work.